+++ src/PhotoGallery/PhotoGallery.Application/Features/Images/Commands/DeleteImage/DeleteImageCommandHandler.cs
@@ -9,23 +9,29 @@ namespace PhotoGallery.Application.Features.Images.Commands.DeleteImage
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IImageService _imageService;
+        private readonly IUserService _userService;
 
-        public DeleteImageCommandHandler(IUnitOfWork unitOfWork, IImageService imageService)
+        public DeleteImageCommandHandler(IUnitOfWork unitOfWork, IImageService imageService, IUserService userService)
         {
             _unitOfWork = unitOfWork;
             _imageService = imageService;
+            _userService = userService;
         }
 
         public async Task Handle(DeleteImageCommand request, CancellationToken cancellationToken)
         {
-            var repo = _unitOfWork.GetRepository<Image>();
-
-            var imageToDelete = await repo.GetByIdAsync(request.Id);
+            var imageToDelete = await _unitOfWork.ImageRepository.GetByIdWithAlbumAsync(request.Id);
 
             if (imageToDelete == null)
             {
                 throw new ArgumentException(
                     $"Image with id: {request.Id} can't be found.", nameof(request.Id));
+            }
+
+            if (imageToDelete.Album.UserId != request.UserId 
+                && !await _userService.IsAdminAsync(request.UserId))
+            {
+                throw new ArgumentException("User have no rights delete image.", nameof(request.UserId));
             }
 
             _imageService.DeleteImage(imageToDelete.FileName);
